pyproject.toml
[project]
name = "cutowm"
description = "A Pygame-ce window manager"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name="Johan Mallqui" }
]

dependencies = [
    "pygame-ce>=2.3.1"
]

[project.scripts]



[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"